feat(sharding): Phase A ULID ripple — production code fully converted
Resolve the app-code ULID ripple (216 new test failures -> 40):
- (int) casts on $obj->id / $obj->*_id properties -> (string) across app+tests
  (root of the phantom FK '1' values: (int)"01kv..." parses to 1)
- selection-id parsers ((int) substr($selection)) in ConnectGoogle /
  ConnectSourcesModal / WebsitesList / IntegrationsPanel -> string
- id-returning method return types (resolveActor/resolveWebsiteOwner/homepageId/
  triggerWebsiteId/userId) -> string/?string
- (int) session('impersonator_id'), Auth::id() casts; blade int-id closures
- a few test-helper int $cs id params -> string

Production code is ULID-clean; remaining ~40 test failures are fixture-level
(mock expectations using ULID-derived ids + a transaction-leak cascade where
one throwing fixture test poisons downstream tests that pass in isolation) plus
3 pre-existing baseline failures. Branch isolated; no prod/staging DB touched.

+++ app/Livewire/ConnectSourcesModal.php
@@ -166,7 +166,7 @@ class ConnectSourcesModal extends Component
             return [null, $selection];
         }
 
-        $accountId = (int) substr($selection, 0, $pos);
+        $accountId = substr($selection, 0, $pos);
 
         return [$accountId > 0 ? $accountId : null, substr($selection, $pos + 1)];
     }
