+++ app/Livewire/Onboarding/ConnectGoogle.php
@@ -251,7 +251,7 @@ class ConnectGoogle extends Component
             return [null, $selection];
         }
 
-        $accountId = (int) substr($selection, 0, $pos);
+        $accountId = substr($selection, 0, $pos);
 
         return [$accountId > 0 ? $accountId : null, substr($selection, $pos + 1)];
     }
